[Refactor] Extract PD disaggregation helpers to pd_utils.py
Move 10 PD-related methods (~260 lines) from OmniBase into a
PDDisaggregationMixin class in a new pd_utils.py file. OmniBase
inherits from the mixin so all self._method() call sites remain
unchanged. This keeps omni.py focused on orchestration logic.
